NotifyMessage: fix instance count
The previous way of counting instances was wrong: it would always stop
at 1 instance maximum, causing issues when instances were destructed. In
particular, when having two notifications active, one of which is a
Yes/No snap decision and the other one is a quick (5 seconds)
notification of a sync failure, both notifications would get dismissed
after just 5 seconds.

This is the last commit needed to completely fix
https://github.com/ubports/ubuntu-touch/issues/1195